Added a battle
Destruction is going on, for two players...

+++ src/main/java/com/capgemini/District.java
@@ -6,7 +6,7 @@ public class District extends Contestant{
     private int drive; //needs to be added to defense
 
     //constructor including super
-    public District(double playerId, boolean isAlive, int health, Gender gender, int attack, int defense, int drive) {
+    public District(int playerId, boolean isAlive, int health, Gender gender, int attack, int defense, int drive) {
         super(playerId, isAlive, health, gender, attack, defense);
         this.drive = drive;
     }
